// src/user/dto/update.dto.ts
export class StoreUpdateDto {
  store_name?: string;

  store_password?: string;

  store_address?: string;

  store_phone_number?: string;

  store_longitude?: string;

  store_latitude?: string;

  store_business_hour?: string;

  store_grade?: number;

  store_image?: string;

  store_profile_image?: string;

  store_closed_day?: string;
}